Resolving MSVC 2017 compiler warnings

+++ src/BabylonCpp/include/babylon/actions/directactions/play_animation_action.h
@@ -25,7 +25,7 @@ public:
    * @param condition defines the trigger related conditions
    */
   PlayAnimationAction(unsigned int triggerOptions, const IAnimatablePtr& target,
-                      int from, int to, bool loop = false,
+                      float from, float to, bool loop = false,
                       Condition* condition = nullptr);
   ~PlayAnimationAction() override;
 
@@ -50,12 +50,12 @@ public:
   /**
    * Where the animation should start (animation frame)
    */
-  int from;
+  float from;
 
   /**
    * Where the animation should stop (animation frame)
    */
-  int to;
+  float to;
 
   /**
    * Define if the animation should loop or stop after the first play.
